Feat : Diary에 Member 정보 추가

+++ src/main/java/com/odos/odos_server_v2/domain/shared/dto/LikeDto.java
@@ -9,5 +9,5 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class LikeDto {
   private boolean likedByMe;
-  private int likeCnt;
+  private long likeCnt;
 }
